pyproject.toml
[project]
name = "left-click-macro"
version = "0.1.0"
description = "Add your description here"
dependencies = [
    "python-uinput>=1.0.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[project.scripts]
"left_click_macro" = "left_click_macro:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "black>=25.1.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/left_click_macro"]
